Moved mapping helpers - Created a new 'helpers' folder in the 'Mapping' component folder - Moved relevant helpers from 'src/Helpers' - Updated paths
