+++ packages/client/src/Components/Mapping/helpers/MapAdjudicatingListener.js
@@ -1,7 +1,7 @@
 import L from 'leaflet'
-import defaultHexStyle from './data/default-hex-style'
+import defaultHexStyle from '../../../Helpers/data/default-hex-style'
 
-export default class MapAdjudicatingPlayerListener {
+export default class MapAdjudicatingListener {
   constructor (map, grid) {
     this.grid = grid
 
@@ -82,9 +82,6 @@ export default class MapAdjudicatingPlayerListener {
             return cell.sea
           } else if (marker.travelMode === 'air') {
             return true
-          } else {
-            console.error('Unexpected terrain type')
-            return false
           }
         })
