+++ packages/client/src/Components/Mapping/helpers/MapAdjudicatingUmpireListener.js
@@ -1,8 +1,8 @@
 import L from 'leaflet'
-import defaultHexStyle from './data/default-hex-style'
-import colorFor from './colorFor'
-import declutterMarkers from './declutterMarkers'
-import plannedStateFor from './plannedStateFor'
+import defaultHexStyle from '../../../Helpers/data/default-hex-style'
+import colorFor from '../../../Helpers/colorFor'
+import declutterMarkers from '../../../Helpers/declutterMarkers'
+import plannedStateFor from '../../../Helpers/plannedStateFor'
 
 export default class MapAdjudicatingListener {
   constructor (map, grid, planningFormCallback) {
@@ -202,7 +202,7 @@ export default class MapAdjudicatingListener {
           const turnMarker = L.marker(marker.coord, {
             draggable: true
           })
-          
+
           // store the force/asset name
           res.name = asset.force + '_' + asset.name
