+++ packages/client/src/Components/Mapping/helpers/MapAdjudicationPendingListener.js
@@ -1,6 +1,6 @@
 import L from 'leaflet'
-import defaultHexStyle from './data/default-hex-style'
-import colorFor from './colorFor'
+import defaultHexStyle from '../../../Helpers/data/default-hex-style'
+import colorFor from '../../../Helpers/colorFor'
 
 export default class MapAdjudicationPendingListener {
   constructor (map, grid, callback) {
