+++ packages/client/src/Components/Mapping/helpers/MapPlanningUmpireListener.js
@@ -1,5 +1,5 @@
 import L from 'leaflet'
-import defaultHexStyle from './data/default-hex-style'
+import defaultHexStyle from '../../../Helpers/data/default-hex-style'
 
 export default class MapPlanningPlayerListener {
   constructor (map, grid) {
