+++ packages/client/src/Components/Mapping/index.jsx
@@ -1,12 +1,12 @@
 import React, { useEffect, useRef, useState } from 'react'
 import L from 'leaflet'
 import GridImplementation from '../../Helpers/GridImplementation'
-import MapAdjudicatingUmpireListener from '../../Helpers/MapAdjudicatingUmpireListener'
-import MapAdjudicatingPlayerListener from '../../Helpers/MapAdjudicatingPlayerListener'
-import MapAdjudicationPendingListener from '../../Helpers/MapAdjudicationPendingListener'
-import MapMarkersControl from '../../Helpers/MapMarkersControl'
-import MapPlanningPlayerListener from '../../Helpers/MapPlanningPlayerListener'
-import MapPlanningUmpireListener from '../../Helpers/MapPlanningUmpireListener'
+import MapAdjudicatingUmpireListener from './helpers/MapAdjudicatingUmpireListener'
+import MapAdjudicatingPlayerListener from './helpers/MapAdjudicatingPlayerListener'
+import MapAdjudicationPendingListener from './helpers/MapAdjudicationPendingListener'
+import MapMarkersControl from './helpers/MapMarkersControl'
+import MapPlanningPlayerListener from './helpers/MapPlanningPlayerListener'
+import MapPlanningUmpireListener from './helpers/MapPlanningUmpireListener'
 import markerFor from '../../Helpers/markerFor'
 import hasPendingForces from '../../Helpers/hasPendingForces'
 import { saveMapMessage } from '../../ActionsAndReducers/playerUi/playerUi_ActionCreators'
@@ -182,7 +182,6 @@ const Mapping = ({ currentTurn, role, currentWargame, selectedForce, allForces, 
         if (myForceRef.current === 'umpire') {
           mapListenerRef.current = new MapPlanningUmpireListener(mapRef.current, gridImplRef.current)
         } else {
-
           mapListenerRef.current = new MapPlanningPlayerListener(mapRef.current, gridImplRef.current, myForceRef.current, currentTurnRef.current, routeComplete)
         }
         break
